requirements.txt
psycopg2-binary==2.9.9
numpy==1.26.4
pgvector==0.2.5
sentence-transformers==2.7.0
